+++ src/scripts/searchcontent/searchcontent.sh
@@ -1,15 +1,30 @@
 #!/usr/bin/env zsh
 
 OUTPUT_FILE="$1"
+searchPaths=($(eval echo "$FZFSEARCH_SEARCH_PATHS"))
 
 source "$(dirname "$0")/../config.sh"
+previewPath="$(dirname "$0")/../preview.sh"
 
+queryFileContent="$(dirname "$0")/query_file_content.sh"
+
+selected=$(
 fzf --phony --query "" \
-  --preview "bat --color=always --plain --highlight-line {2} {1} 2>/dev/null || true" \
+  --preview "source $previewPath && preview {1} {2}" \
   --delimiter ':' \
   --preview-window "+{2}-10" \
-  --bind "change:reload:(rg --hidden --glob $RIPGREP_GLOB -n {q} || true)" \
-  --bind "start:reload:(rg --hidden --glob $RIPGREP_GLOB -n {q} || true)" \
+  --bind "change:reload:(source $queryFileContent && query_file_content {q})" \
+  --bind "start:reload:(source $queryFileContent && query_file_content {q})" \
   --bind "$KEYMAPPING" \
   --layout "$LAYOUT" \
-  --multi > "${OUTPUT_FILE}";
+  --multi
+)
+
+echo "$selected" | while read -r file; do
+  for dir in "${searchPaths[@]}"; do
+    base=$(basename "$dir")
+    if [[ "$file" == "$base"* ]]; then
+      (echo "$file" | sed "s|^$base|$dir|")
+    fi
+  done
+done > "$OUTPUT_FILE"