+++ src/scripts/searchfile/fzf_reload_by_query.sh
@@ -1,36 +1,42 @@
 #!/usr/bin/env zsh
-
 fzf_reload_by_query() {
   q="$1"
+  searchPaths=($(eval echo "$FZFSEARCH_SEARCH_PATHS"))
+  typeset -A dir_map
+  for dir in "${searchPaths[@]}"; do
+    dir_map[${dir:t}]="$dir"
+  done
 
   if [ -z "$q" ]; then
-    rg --files
+    print -rNC1 -- "${searchPaths[@]}" | xargs -0 -P 0 -I {} zsh -c '
+      dir="{}"
+      base="${dir:t}"
+      dir="${dir%/}" 
+      glob="$1"
+      rg --hidden --glob "$RIPGREP_GLOB" --files "$dir" | sed "s|^$dir|$base|"
+    ' _ "$RIPGREP_GLOB"
     return
   fi
 
   name=$(printf "%s" "$q:" | cut -d: -f1)
   line=$(printf "%s" "$q:" | cut -d: -f2)
-  res=$(rg --hidden --glob $RIPGREP_GLOB --files | fzf -f "$name")
+  print -rNC1 -- "${searchPaths[@]}" | xargs -0 -P 0 -I {} zsh -c '
+    dir="$1"
+    glob="$2"
+    name="$3"
+    line="$4"
 
-  if [ -z "$res" ]; then
-    return
-  fi
+    base="${dir:t}"
+    dir="${dir%/}"
 
-  if [ -z "$line" ]; then
-    if [[ "$q" == *:* ]]; then
-      echo "$res" | xargs -I{} echo {}:
-    else
-      echo "$res"
+    result=$(rg --hidden --glob "$glob" --files "$dir" | sed "s|^$dir|$base|" | fzf -f "$name")
+
+    if [ -n "$result" ]; then
+      if [ -n "$line" ]; then
+        echo "${result}" | sed "s|$|:$line|"
+      else
+        echo "$result"
+      fi
     fi
-    return
-  fi
-
-  echo "$res" | while read -r file; do
-    max_line=$(wc -l < "$file" | tr -d " ")
-    if [ "$line" -gt "$max_line" ]; then
-      echo "$file:$max_line"
-    else
-      echo "$file:$line"
-    fi
-  done
+  ' _ {} "$RIPGREP_GLOB" "$name" "$line"
 }