+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [v0.0.6] - 2025-12-15
+
+- Enhancement: Support search with VSCode workspace folders.
+- Enhancement: add shortcut ctrl-t to open repository in current workspace.
+- Change: Updated use --smart-case flag by default with rg.
+- Improvement: use xargs multi-processing to speed up search.
+
 ## [v0.0.4] - 2025-12-10
 
 - Enhancement: Support for searching within specific repositories by adding a new configuration option "repositoryPath".